Fixed `TypeError: int() argument must be a string, a bytes-like object or a real number, not 'list'
I've fixed `TypeError: int() argument must be a string, a bytes-like object or a real number, not 'list'` error. It was occuring due to my last commit, i just forgot that the channel ids was in single list, and i add bot owner fsub channel id in it as second value in that, that's why it was occuring. Now i added split every channel id and after that it will add owner fsub channel id in it and do left work.